feat(#133): unify activity bar builder into shared render::build_activity_bar
Both GTK and TUI had near-identical ~80-line functions that each built a
`quadraui::ActivityBar` primitive from engine state. This extracts the shared
logic into `render::build_activity_bar(engine, theme, include_hamburger,
active_ext_panel)` in render.rs — the same adapter pattern used by all other
sidebar primitives (sc_sidebar_panel, populate_explorer_tree_controller, etc.).

Both backends now delegate to this single function:
- GTK: `include_hamburger=false` (menu bar is a native GTK widget),
       passes `engine.ext_panel_active.as_deref()`
- TUI: `include_hamburger=true`  (no native menu bar; hamburger at index 0),
       passes `sidebar.ext_panel_name.as_deref()`

Icons now use `icon.s()` which respects `settings.use_nerd_fonts` — fixing a
latent bug where the GTK builder hardcoded `.nerd` regardless of that setting.
Minor cosmetic change: GTK search icon switches from SEARCH_COD (U+EA6D) to
SEARCH (U+F002); both are magnifying-glass glyphs, fallback "/" is unchanged.
Tooltips now consistently include keyboard shortcuts on both backends (GTK
already showed them; TUI ignores tooltips on activity bar items).

Also fixes a misplaced doc comment: `/// Map GDK key names…` was adjacent to
the removed function's doc block and is now properly placed above
`fn map_gtk_key_name`.

net: -80 lines of duplicated backend-specific builder code.
